# Improve launch performance of kernels some more